fixed debug out for 64 bit systems
git-svn-id: https://lmms.svn.sf.net/svnroot/lmms/trunk/lmms@1753 0778d3d1-df1d-0410-868b-ea421aaaa00d

+++ src/core/mixer.cpp
@@ -215,9 +215,10 @@ private:
 				(fx_ch_t) it->effectChannelJob );
 							break;
 						default:
-fprintf( stderr, "invalid job item type %d at %d in jobqueue(%d:%d)\n",
-		(int) it->type, (int) it, (int) s_jobQueue.items.begin(),
-							(int) end_it );
+fprintf( stderr, "invalid job item type %d at %ld in jobqueue(%ld:%ld)\n",
+		(int) it->type, (long int) it,
+					(long int) s_jobQueue.items.begin(),
+							(long int) end_it );
 							break;
 					}
 				}
